# docs/swagger.yaml
definitions:
  domain.CreateTaskRequest:
    properties:
      description:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.TaskStatus'
        enum:
        - todo
        - in_progress
        - done
      title:
        type: string
    required:
    - description
    - status
    - title
    type: object
  domain.Message:
    properties:
      code:
        type: integer
      description:
        type: string
    type: object
  domain.Task:
    properties:
      description:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/domain.TaskStatus'
      title:
        type: string
    type: object
  domain.TaskStatus:
    enum:
    - todo
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - TaskStatusTodo
    - TaskStatusInProgress
    - TaskStatusDone
  domain.UpdateStatusTaskRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.TaskStatus'
        enum:
        - todo
        - in_progress
        - done
    required:
    - status
    type: object
  domain.UpdateTaskRequest:
    properties:
      description:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.TaskStatus'
        enum:
        - todo
        - in_progress
        - done
      title:
        type: string
    required:
    - description
    - status
    - title
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: get all task
      parameters:
      - in: query
        name: description
        type: string
      - in: query
        name: id
        type: integer
      - enum:
        - todo
        - in_progress
        - done
        in: query
        name: status
        type: string
        x-enum-varnames:
        - TaskStatusTodo
        - TaskStatusInProgress
        - TaskStatusDone
      - in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: GetAll
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: create task
      parameters:
      - description: input create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Create
      tags:
      - Task
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: delete task
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Delete
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: get task by id
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: GetOneByID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: update task
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      - description: input update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Update
      tags:
      - Task
  /tasks/{id}/status:
    patch:
      consumes:
      - application/json
      description: update task status
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      - description: input update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateStatusTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: UpdateStatus
      tags:
      - Task
swagger: "2.0"